feat: Сделал request и response рекордами
