+++ Poolit/Models/Requests/RegisterRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Poolit.Models.Requests;
+
+public class RegisterRequest
+{
+    public string Login { get; set; }
+    public string Password { get; set; }
+}
