+++ Poolit/Models/User.cs
@@ -1,0 +1,10 @@
+﻿namespace Poolit.Models;
+
+[Serializable]
+public class User
+{
+    public int Id { get; set; }
+    public required string Login { get; set; }
+    public string HashedPassword { get; set; }
+    public string Token { get; set; }
+}
