--- Poolit/Models/User/Requests/RegisterRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace Poolit.Models.User.Requests
-{
-    public class RegisterRequest
-    {
-        public string Login { get; set; }
-        public string Password { get; set; }
-    }
-}
