--- Poolit/Models/User/User.cs
@@ -1,11 +1,0 @@
-﻿namespace Poolit.Models.User
-{
-    [Serializable]
-    public class User
-    {
-        public int Id { get; set; }
-        public required string Login { get; set; }
-        public string HashedPassword { get; set; }
-        public string Token { get; set; }
-    }
-}
